requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# RAG & Vector Database (for future phases)
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Local LLM (Ollama)
ollama>=0.1.0

# Web Scraping (for Phase 3)
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0
cloudscraper>=1.2.71
lxml>=4.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Utilities
tqdm>=4.66.0
python-json-logger>=2.0.7

# UI - Streamlit Interface
streamlit>=1.31.0

